@@ -7,6 +7,7 @@
 #define __ALPHA_PROVIDER_IMPL_H
 
 #include "alpha/ResourceInterface.hpp"
+#include "alpha/BulkLocation.hpp"
 
 #include <thallium.hpp>
 #include <thallium/serialization/stl/string.hpp>
@@ -15,6 +16,7 @@
 #include <nlohmann/json.hpp>
 #include <spdlog/spdlog.h>
 
+#include <format>
 #include <tuple>
 
 namespace alpha {
@@ -28,11 +30,11 @@ class ProviderImpl : public tl::provider<ProviderImpl> {
 
     using json = nlohmann::json;
 
-    #define DEF_LOGGING_FUNCTION(__name__)                         \
-    template<typename ... Args>                                    \
-    void __name__(Args&&... args) {                                \
-        auto msg = fmt::format(std::forward<Args>(args)...);       \
-        spdlog::__name__("[alpha:{}] {}", get_provider_id(), msg); \
+    #define DEF_LOGGING_FUNCTION(__name__)                           \
+    template<typename ... Args>                                      \
+    void __name__(std::format_string<Args...> fmt, Args&&... args) { \
+        auto msg = std::format(fmt, std::forward<Args>(args)...);    \
+        spdlog::__name__("[alpha:{}] {}", get_provider_id(), msg);   \
     }
 
     DEF_LOGGING_FUNCTION(trace)
@@ -50,6 +52,7 @@ class ProviderImpl : public tl::provider<ProviderImpl> {
     tl::pool             m_pool;
     // Client RPC
     tl::auto_remote_procedure m_compute_sum;
+    tl::auto_remote_procedure m_compute_sum_bulk;
     // FIXME: other RPCs go here ...
     // ResourceInterfaces
     std::shared_ptr<ResourceInterface> m_backend;
@@ -59,6 +62,7 @@ class ProviderImpl : public tl::provider<ProviderImpl> {
     , m_engine(engine)
     , m_pool(pool)
     , m_compute_sum(define("alpha_compute_sum",  &ProviderImpl::computeSumRPC, pool))
+    , m_compute_sum_bulk(define("alpha_compute_sum_bulk",  &ProviderImpl::computeSumBulkRPC, pool))
     {
         trace("Registered provider with id {}", get_provider_id());
         json json_config;
@@ -135,13 +139,52 @@ class ProviderImpl : public tl::provider<ProviderImpl> {
         trace("Received computeSum request");
         Result<int32_t> result;
         tl::auto_respond<decltype(result)> response{req, result};
-        if(!m_backend) {
-            result.success() = false;
-            result.error() = "Provider has no resource attached";
-        } else {
-            result = m_backend->computeSum(x, y);
-        }
+        result = m_backend->computeSum(x, y);
         trace("Successfully executed computeSum");
+    }
+
+    void computeSumBulkRPC(const tl::request& req,
+                           BulkLocation remote_x, BulkLocation remote_y,
+                           BulkLocation remote_result) {
+        trace("Received computeSumBulk request");
+        Result<bool> result;
+        tl::auto_respond<decltype(result)> response{req, result};
+        auto n = remote_x.size / sizeof(int32_t);
+        try {
+
+            auto x_endpoint      = m_engine.lookup(remote_x.address);
+            auto y_endpoint      = m_engine.lookup(remote_y.address);
+            auto result_endpoint = m_engine.lookup(remote_result.address);
+
+            std::vector<int32_t> local_x(n);
+            std::vector<int32_t> local_y(n);
+            std::vector<int32_t> local_result(n);
+
+            auto local_x_bulk = m_engine.expose(
+                {{(void*)local_x.data(), local_x.size()*sizeof(int32_t)}},
+                tl::bulk_mode::write_only);
+            auto local_y_bulk = m_engine.expose(
+                {{(void*)local_y.data(), local_y.size()*sizeof(int32_t)}},
+                tl::bulk_mode::write_only);
+            auto local_result_bulk = m_engine.expose(
+                {{(void*)local_result.data(), local_result.size()*sizeof(int32_t)}},
+                tl::bulk_mode::read_only);
+
+            local_x_bulk << remote_x.bulk(remote_x.offset, remote_x.size).on(x_endpoint);
+            local_y_bulk << remote_y.bulk(remote_y.offset, remote_y.size).on(y_endpoint);
+
+            for(size_t i = 0; i < n; ++i) {
+                auto r = m_backend->computeSum(local_x[i], local_y[i]);
+                r.check();
+                local_result[i] = r.value();
+            }
+
+            local_result_bulk >> remote_result.bulk(remote_result.offset, remote_result.size).on(result_endpoint);
+        } catch(const std::exception& ex) {
+            result.error() = ex.what();
+            result.success() = false;
+        }
+        trace("Successfully executed computeSumBulk");
     }
 
 };
